disable new behavior warnings for all of atupone's recent changes fix double to float cast warning.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@22257 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/playing.cxx
@@ -6637,7 +6637,7 @@ static void		playingLoop()
 	nextTime += 1.0f / fpsLimit;
 	float remaining;
         while (1) {
-	  remaining = nextTime - TimeKeeper::getCurrent();
+	  remaining = (float)(nextTime - TimeKeeper::getCurrent());
 	  if (remaining > 1.0f)
 	    break;
 	  if (remaining <= 0.0f)
